+++ arch/xtensa/src/lib.rs
@@ -23,6 +23,7 @@ pub mod appcpu;
 pub mod critical_section;
 pub mod mpu;
 pub mod registers;
+pub mod smp;
 pub mod tick;
 
 pub use critical_section::{with as cs_with, XtensaCriticalSection, XtensaCsToken};